progressive changes

+++ src/contexts/AuthContext/index.tsx
@@ -2,32 +2,15 @@ import React, { createContext, useContext, useEffect, useState } from 'react';
 import { AxiosError } from 'axios';
 import toastMsg, { ToastType } from '../../utils/toasts/toastMsg';
 import setTokenStorage from '../../utils/tokens/setTokenStorage';
-import { setAxiosAuth } from '../../services/HTTPClient';
+import HTTPClient, { setAxiosAuth } from '../../services/HTTPClient';
 import { UsersService } from '../../services/server/users/user.service';
 import { IUser } from '../../interfaces/IUsers';
-
-interface IContextUser {
-  _id: string;
-  name: string;
-  lastName: string;
-  email: string;
-  subscriptionsId?: '';
-  supportsId?: '';
-  casterId?: '';
-  exp: number;
-}
-
-interface IContextLogin {
-  data: IUser;
-  authorization: string;
-}
 
 interface AuthContextData {
   logged: boolean;
   user: IUser;
-  Login(loginData: { email: string; password: string }): Promise<void>;
+  Login(loginData: { email: string; password: string }): Promise<boolean>;
   Logout(): void;
-  checkToken(): boolean;
   updateUserName(newName: string): void;
 }
 
@@ -55,16 +38,24 @@ export const AuthProvider = ({
   children: React.ReactElement;
 }): React.ReactElement => {
   const [user, setUser] = useState<IUser>(emptyUser);
+  const [logged, setLogged] = useState<boolean>(false);
 
-  const Login = async (loginData: { email: string; password: string }): Promise<void> => {
+  const Login = async (loginData: { email: string; password: string }): Promise<boolean> => {
     try {
-      console.log('useAuth login');
-      const { authorization, data } = await UsersService.login(loginData);
+      const loginResp = await UsersService.login(loginData);
 
-      console.log('Token data:', data);
+      if (typeof loginResp === 'boolean') {
+        alert('Invalid login');
+        return false;
+      }
+
+      const { authorization, data } = loginResp;
       localStorage.clear();
 
+      HTTPClient.api.defaults.headers.common.authorization = authorization;
+
       setUser(data);
+      setLogged(true);
 
       setTokenStorage('authorization', authorization);
       localStorage.setItem(
@@ -77,29 +68,19 @@ export const AuthProvider = ({
           subscriptionsId: data.subscriptionsId,
           supportsId: data.supportsId,
           casterId: data.casterId,
-          // exp: data.exp,
         })
       );
+
+      return true;
     } catch (error) {
-      toastMsg(ToastType.Error, (error as AxiosError).message || 'Internal Server Error!');
+      return false;
     }
   };
 
   const Logout = (): void => {
     localStorage.clear();
     setUser(emptyUser);
-  };
-
-  const checkToken = (): boolean => {
-    let isValid = false;
-    // if (user.exp) {
-    //   if (new Date(user.exp * 1000) < new Date()) {
-    //     Logout();
-    //     toastMsg(ToastType.Warning, 'Sua sessão expirou.');
-    //   } else isValid = true;
-    // }
-
-    return isValid;
+    setLogged(false);
   };
 
   const updateUserName = (newName: string): void => {
@@ -109,7 +90,7 @@ export const AuthProvider = ({
   useEffect(() => {
     let isCleaning = false;
     if (!isCleaning) {
-      checkToken();
+      //!Do something
     }
     return () => {
       isCleaning = true;
@@ -126,6 +107,7 @@ export const AuthProvider = ({
 
       if (localToken && localUser) {
         setUser(JSON.parse(localUser));
+        setLogged(true);
       }
     }
     return () => {
@@ -134,9 +116,7 @@ export const AuthProvider = ({
   }, []);
 
   return (
-    <AuthContext.Provider
-      value={{ logged: !!user.name, user, Login, Logout, checkToken, updateUserName }}
-    >
+    <AuthContext.Provider value={{ logged, user, Login, Logout, updateUserName }}>
       {children}
     </AuthContext.Provider>
   );
